feat(.service): 文件上传ddl

+++ llmops-common/src/main/java/com/emcikem/llm/common/vo/file/UploadFileVO.java
@@ -19,7 +19,7 @@ public class UploadFileVO {
 
     private String key;
 
-    private Long size;
+    private Integer size;
 
     private String extension;
